+++ src/main/java/kr/kro/airbob/domain/discountPolicy/entity/DiscountPolicy.java
@@ -1,4 +1,4 @@
-package kr.kro.airbob.domain.discountPolicy;
+package kr.kro.airbob.domain.discountPolicy.entity;
 
 import java.time.LocalDateTime;
