+++ src/test/java/kr/kro/airbob/domain/discountPolicy/DiscountPolicyServiceTest.java
@@ -1,6 +1,6 @@
-package kr.kro.airbob.domain.discountPolicy.service;
+package kr.kro.airbob.domain.discountPolicy;
 
-import kr.kro.airbob.domain.discountPolicy.DiscountPolicy;
+import kr.kro.airbob.domain.discountPolicy.entity.DiscountPolicy;
 import kr.kro.airbob.domain.discountPolicy.common.DiscountType;
 import kr.kro.airbob.domain.discountPolicy.common.PromotionType;
 import kr.kro.airbob.domain.discountPolicy.dto.request.DiscountPolicyCreateDto;
